checked when we have already volumes

+++ web/start.sh
@@ -10,42 +10,27 @@ mkdir -p /var/lib/drupal-private
 chown -R www-data /var/lib/drupal-private
 
 GIT_REPO="https://github.com/szmediathek/szmediathek.git"
-DATABASE_REPO="git@github.com:szmediathek/databases.git"
-#MYSQL_HOST=""
-#MYSQL_USER=""
-#MYSQL_DATABASE=""
-#MYSQL_ROOT_PASSWORD=""
-MYSQL_PASSWORD="1"
 
-#todo: handle existing repository
 
-echo "02. cloning repo"
-cd /var/www
-rm -rf html
-#todo: use ssh + myphrase
-git clone ${GIT_REPO} html
-cd html
-git checkout stage
+if [ ! -f /var/www/sites/default/settings.php ] ; then
 
-#todo: use ssh + myphrase
-#git clone ${DATABASE_REPO} db
-GIT_SSH=/gitwrap.sh git clone ${DATABASE_REPO} db
-if ! [ -n "$SSH_AUTH_SOCK" ] || 
-  ! { ssh-add -l &>/dev/null; rc=$?; [ "$rc" -eq 0 ] || [ "$rc" -eq 1 ];}; then
-    echo "Starting agent..."
-    eval "$(ssh-agent -s)"
-fi
+    echo "02. cloning repo"
+    cd /var/www
+    rm -rf html
+    #todo: use ssh + myphrase
+    git clone ${GIT_REPO} html
+    cd html
+    git checkout stage
 
-mkdir -p sites/default/files && chmod 755 sites/default && chown -R www-data:www-data sites/default/files;
+    mkdir -p sites/default/files && chmod 755 sites/default && chown -R www-data:www-data sites/default/files;
 
-echo "03. setting database"
-mv /tmp/settings.php sites/default/settings.php
-sed -i "s/placeholder_PWD/${MYSQL_ROOT_PASSWORD}/g" sites/default/settings.php
-sed -i "s/placeholder_DB/${MYSQL_DATABASE}/g" sites/default/settings.php
-sed -i "s/placeholder_USER/${MYSQL_USER}/g" sites/default/settings.php
-sed -i "s/placeholder_HOST/${MYSQL_HOST}/g" sites/default/settings.php
+    echo "03. setting database"
+    mv /tmp/settings.php sites/default/settings.php
+    sed -i "s/placeholder_PWD/${MYSQL_ROOT_PASSWORD}/g" sites/default/settings.php
+    sed -i "s/placeholder_DB/${MYSQL_DATABASE}/g" sites/default/settings.php
+    sed -i "s/placeholder_USER/${MYSQL_USER}/g" sites/default/settings.php
+    sed -i "s/placeholder_HOST/${MYSQL_HOST}/g" sites/default/settings.php
 
-if [[ ${LOCAL_MYSQL} ]]; then
     mysqladmin -u root password $MYSQL_ROOT_PASSWORD 
     #echo "CREATE DATABASE $MYSQL_DATABASE; GRANT ALL PRIVILEGES ON ${MYSQL_DATABASE}.* TO $MYSQL_USER@'localhost' IDENTIFIED BY '$MYSQL_PASSWORD'; FLUSH PRIVILEGES;"
     mysql -uroot -p$MYSQL_ROOT_PASSWORD -e "CREATE DATABASE $MYSQL_DATABASE; GRANT ALL PRIVILEGES ON ${MYSQL_DATABASE}.* TO $MYSQL_USER@'localhost' IDENTIFIED BY '$MYSQL_PASSWORD'; FLUSH PRIVILEGES;"
@@ -53,33 +38,52 @@ if [[ ${LOCAL_MYSQL} ]]; then
     mv /root/.my.cnf.sample /root/.my.cnf
     sed -i "s/ADDED_BY_START.SH/$MYSQL_ROOT_PASSWORD/" /root/.my.cnf 
 
-    #get database - maybe git
-    #copy database
-    #clear cache
-    #delete repo git
+    if [[ "${MYSQL_HOST}" = "mysql" ]]; then        
+        DATABASE_REPO="https://${GIT_USER}:${GIT_PASSWORD}@github.com/szmediathek/databases.git"
+        cd /var/www/html
+        git clone ${DATABASE_REPO} db
+        cd db
+        echo "02. unzip sql file"
+        gunzip -c ${FILENAME} > /tmp/db1.sql
+        cd ..
+        #drush sql-drop #check when
+        echo "02. import database"
+        drush sql-cli < /tmp/db1.sql
+        rm /tmp/db1.sql        
+    fi
+
+else
+    echo "02. pulling repo"
+    cd /var/www/html
+    git pull
+
+    if [[ "${UPDATE_DB}" = "1" ]]; then
+        DATABASE_REPO="https://${GIT_USER}:${GIT_PASSWORD}@github.com/szmediathek/databases.git"
+        cd /var/www/html
+        if [ ! -d /var/www/html/db ] ; then
+            git clone ${DATABASE_REPO} db
+        else
+            cd db
+            git pull
+            cd ..
+        fi
+        cd db
+        echo "02. unzip sql file"
+        gunzip -c ${FILENAME} > /tmp/db1.sql
+        cd ..
+        #drush sql-drop #check when
+        echo "02. import database"
+        drush sql-cli < /tmp/db1.sql
+        rm /tmp/db1.sql
+    fi
 fi
 
-echo "4. setting up solr"
-#cd /opt
-#wget http://archive.apache.org/dist/lucene/solr/4.7.2/solr-4.7.2.tgz
-#tar -xvf solr-4.7.2.tgz
-#cp -R solr-4.7.2/example /opt/solr
-#cd /opt/solr
-#java -jar start.jar
-#sudo cat /tmp/jetty > /etc/default/jetty
-#sudo cat /tmp/jetty-logging.xml > /opt/solr/etc/jetty-logging.xml
-#sudo useradd -d /opt/solr -s /sbin/false solr
-#sudo chown solr:solr -R /opt/solr
-#sudo wget -O /etc/init.d/jetty http://git.eclipse.org/c/jetty/org.eclipse.jetty.project.git/plain/jetty-distribution/src/main/resources/bin/jetty.sh
-#sudo chmod a+x /etc/init.d/jetty
-#sudo update-rc.d jetty defaults
-#sudo /etc/init.d/jetty start
-#use docker composer
-#will not work because we have another port
-   
-# Create log that can be written to in the running container, and visible in the
-# docker log (stdout) and thus the webfact UI.
-# todo: could this be done by supervisord, but it must send the tail to stdout?
+if [ "x$MYSQL_HOST" == 'xlocalhost' ] ; then
+    # Stop mysql, will be restarted by supervisor below
+    killall mysqld
+    sleep 5s
+fi
+
 webfactlog=/tmp/webfact.log;
 echo "`date '+%Y-%m-%d %H:%M'` Create new $webfactlog" > $webfactlog
 tail -f $webfactlog &
